rename generic file
