+++ ucloud/utest/driver/env.go
@@ -2,7 +2,7 @@ package driver
 
 import "os"
 
-const ACC_ENV_KEY = "USDK_ACC"
+const ACC_ENV_KEY = "USDKACC"
 const ACC_SKIP_REASON = "skip test for non-acc environment"
 
 // check test env is acceptance testing or not
